+++ Source/PLCrashAsyncMachOImageTests.m
@@ -28,7 +28,6 @@
 
 #import "GTMSenTestCase.h"
 
-#import "PLCrashImageAnnotation.h"
 #import "PLCrashAsyncMachOImage.h"
 
 #import <dlfcn.h>
@@ -36,10 +35,6 @@
 #import <mach-o/getsect.h>
 #import <objc/runtime.h>
 #import <execinfo.h>
-
-/** Image-wide annotation used in our unit tests below. Note that there can only be one such attribute set
-  * in a given Mach-O image; our use of it here precludes use in other tests. */
-PLCrashImageAnnotation ImageGlobalAnnotation PLCRASH_IMAGE_ANNOTATION_ATTRIBUTE;
 
 @interface PLCrashAsyncMachOImageTests : SenTestCase {
     /** The image containing our class. */
@@ -51,11 +46,6 @@ PLCrashImageAnnotation ImageGlobalAnnotation PLCRASH_IMAGE_ANNOTATION_ATTRIBUTE;
 @implementation PLCrashAsyncMachOImageTests
 
 - (void) setUp {
-    /* Zero our image annotation */
-    ImageGlobalAnnotation.version = 0;
-    ImageGlobalAnnotation.data = NULL;
-    ImageGlobalAnnotation.data_size = 0;
-
     /* Fetch our containing image's dyld info */
     Dl_info info;
     STAssertTrue(dladdr([self class], &info) > 0, @"Could not fetch dyld info for %p", [self class]);
@@ -433,81 +423,6 @@ static void testFindSymbol_cb (pl_vm_address_t address, const char *name, void *
 
     /* Compare the results */
     STAssertEquals((pl_vm_address_t) localIMP, pc, @"Returned incorrect symbol address");
-}
-
-/**
- * Test crash info lookup for images which provide annotation information via
- * the __DATA,__plcrash_info section.
- */
-- (void) testFindAnnotation {
-    plcrash_async_mobject_t annotation;
-
-    /* Try to find the annotation; this should fail if the data_size is 0 */
-    ImageGlobalAnnotation.data_size = 0;
-    plcrash_error_t ret = plcrash_async_macho_find_annotation(&_image, &annotation);
-    STAssertEquals(ret, PLCRASH_ENOTFOUND, @"Somehow found image annotation");
-
-
-    /* Initialize the annotation and try finding it again */
-    const char *test_data = "hello, world";
-    ImageGlobalAnnotation.version = 0;
-    ImageGlobalAnnotation.data = test_data;
-    ImageGlobalAnnotation.data_size = strlen(test_data);
-
-    ret = plcrash_async_macho_find_annotation(&_image, &annotation);
-    STAssertEquals(ret, PLCRASH_ESUCCESS, @"Failed to get image annotation");
-    STAssertEquals(annotation.length, (pl_vm_size_t)ImageGlobalAnnotation.data_size, @"Annotation length is wrong");
-
-
-    /* Verify that the fetched annotation data is correct */
-    void *data = plcrash_async_mobject_remap_address(&annotation, annotation.task_address, 0, annotation.length);
-    STAssertNotNULL(data, @"Failed to remap annotation string");
-
-    if (data) {
-        /* If the last test failed, don't come here and crash. */
-        STAssertEquals(strncmp((const void *) ImageGlobalAnnotation.data, data, annotation.length), 0, @"Annotation data is wrong");
-    }
-
-
-    /* Clean up */
-    plcrash_async_mobject_free(&annotation);
-}
-
-/**
- * Test crash info lookup to make sure it handles invalid data correctly.
- */
-- (void) testFindAnnotationInvalid {
-    plcrash_async_mobject_t annotation;
-    
-    /* Search on a NULL Mach-O image should return EINVAL */
-    plcrash_error_t ret = plcrash_async_macho_find_annotation(NULL, &annotation);
-    STAssertEquals(ret, PLCRASH_EINVAL, @"Didn't return invalid status");
-
-
-    /* Search with a NULL annotation argument should return EINVAL */
-    ret = plcrash_async_macho_find_annotation(&_image, NULL);
-    STAssertEquals(ret, PLCRASH_EINVAL, @"Didn't return invalid status");
-
-
-    /* Search with all NULL arguments should return EINVAL */
-    ret = plcrash_async_macho_find_annotation(NULL, NULL);
-    STAssertEquals(ret, PLCRASH_EINVAL, @"Didn't return invalid status");
-
-
-    /* Search on an annotation with a NULL data pointer should return PLCRASH_EINVALID_DATA */
-    ImageGlobalAnnotation.version = 0;
-    ImageGlobalAnnotation.data = NULL;
-    ImageGlobalAnnotation.data_size = 42;
-    ret = plcrash_async_macho_find_annotation(&_image, &annotation);
-    STAssertEquals(ret, PLCRASH_EINVALID_DATA, @"Didn't return invalid status");
-
-
-    /* Search on an annotation with an unsupported version (1) should return PLCRASH_EINVALID_DATA */
-    ImageGlobalAnnotation.version = 1;
-    ImageGlobalAnnotation.data = "123";
-    ImageGlobalAnnotation.data_size = 3;
-    ret = plcrash_async_macho_find_annotation(&_image, &annotation);
-    STAssertEquals(ret, PLCRASH_EINVALID_DATA, @"Didn't check version");
 }
 
 @end
